enhance: implement euler tour (Node), lca Tarjan offline

+++ src/kgmk/dsa/number_theory/prime_factorize/naive/jit.py
@@ -31,7 +31,7 @@ def prime_factorize_factorial(
   n: int,
 ) -> np.array:
   prime, cnt = [], []
-  idx = np.full(n + 1, -1, dtype=np.int32)
+  idx = np.full(n + 1, -1, dtype=np.int64)
   for i in range(n + 1):
     for p, c in prime_factorize(i):
       i = idx[p]
